--- src/app/(providers)/(root)/[workspaceId]/(homeLayout)/home/page.tsx
@@ -1,38 +1,0 @@
-'use client';
-import useWorkspaceUser from '@/hooks/useWorkspaceUser';
-import useUserStore from '@/store/userStore';
-import { useShallow } from 'zustand/react/shallow';
-import Header from './_components/Header';
-import HomeMemberCard from './_components/HomeMemberCard';
-import MemberExistComponent from './_components/MemberExistComponent';
-import MemberNotExistComponent from './_components/MemberNotExistComponent';
-
-const Homepage = () => {
-  const userInfo = {
-    name: '이름',
-    position: 'Position',
-    status: 'Status'
-  };
-  const workspaceUserList = ['user'];
-  const { workspaceUserId, workspaceList } = useUserStore(
-    useShallow((state) => ({
-      workspaceUserId: state.workspaceUserId,
-      workspaceList: state.workspaceList
-    }))
-  );
-  const { workspaceUser } = useWorkspaceUser();
-
-  if (!workspaceUser) return;
-  return (
-    <div>
-      <Header />
-      <main className="px-[16px] mt-[26px]">
-        <HomeMemberCard name={workspaceUser.name} position={workspaceUser.position} status={workspaceUser.state} />
-
-        {workspaceUserList.length === 0 ? <MemberNotExistComponent /> : <MemberExistComponent />}
-      </main>
-    </div>
-  );
-};
-
-export default Homepage;
