+++ src/app/(providers)/(root)/[workspaceId]/_components/InviteCodeButton/InviteCodeButton.tsx
