+++ src/app/(providers)/(root)/[workspaceId]/_components/MemberNotExistComponent/index.ts
